chore: react query and mutation implemented instead of custom network layer

+++ packages/frontend/src/network/useDeleteLink.network.ts
@@ -1,11 +1,17 @@
 import axios from 'axios';
+import { useMutation } from 'react-query';
 
 import { ApiPaths } from '../config/paths.config';
 import { joinPath } from '../utils/path';
-import { useNetwork } from '../utils/useNetwork';
 
-export function useDeleteLink(id: string) {
-  return useNetwork(() => {
-    return axios.delete(joinPath(ApiPaths.LINK, id));
-  });
+export function useDeleteLink(id: string | undefined, onSuccess: () => void) {
+  return useMutation(
+    ['deleteLink', id],
+    async () => {
+      if (!id) return;
+      const response = await axios.delete(joinPath(ApiPaths.LINK, id));
+      return response.data;
+    },
+    { onSuccess }
+  );
 }
